+++ src/com/koxudaxi/pydantic/PydanticParametersProvider.kt
@@ -17,7 +17,7 @@ class PydanticParametersProvider : PyDataclassParametersProvider {
     override fun getDecoratorAndTypeAndParameters(project: Project): Triple<QualifiedName, PyDataclassParameters.Type, List<PyCallableParameter>> {
         val generator = PyElementGenerator.getInstance(project)
         val parameters = mutableListOf(PyCallableParameterImpl.psi(generator.createSingleStarParameter()))
-        parameters.addAll(DATACLASS_ARGUMENTS.map { name -> PyCallableParameterImpl.nonPsi(name, null, null) })
+        parameters.addAll(DATACLASS_ARGUMENTS.map { name -> PyCallableParameterImpl.nonPsi(name, null) })
         return Triple(DATACLASS_QUALIFIED_NAME, PydanticType, parameters)
     }
 
@@ -65,6 +65,7 @@ class PydanticParametersProvider : PyDataclassParametersProvider {
             frozen = false,
             matchArgs = true,
             kwOnly = false,
+            slots = false,
             initArgument = null,
             reprArgument = null,
             eqArgument = null,
@@ -73,10 +74,16 @@ class PydanticParametersProvider : PyDataclassParametersProvider {
             frozenArgument = null,
             matchArgsArgument = null,
             kwOnlyArgument = null,
-            type = PydanticDataclassBypassType,
-            slots = false,
             slotsArgument = null,
+            type = PydanticDataclassBypassType,
             others = emptyMap(),
+            fieldSpecifiers = emptyList(),
+            populateByName = null,
+            validateByName = null,
+            validateByAlias = null,
+            populateByNameArgument = null,
+            validateByAliasArgument = null,
+            validateByNameArgument = null,
         )
     }
 }