+++ testData/completion/subscriptionNonReference.py
@@ -1,0 +1,5 @@
+
+from pydantic import BaseModel
+
+
+(A if True else B)[str]().<caret>