+++ testData/inspectionv18/pydanticDataclassCallable.py
@@ -1,0 +1,10 @@
+from pydantic.dataclasses import dataclass
+
+
+@dataclass
+class A:
+    a: str
+    b: int
+
+
+A(a="x", b=1)