+++ testData/typedvalidatormethodhandler/doesNotLoadPydanticVersionDuringWriteAction.py
@@ -1,0 +1,1 @@
+def plain_function<caret>