+++ testData/typedvalidatormethodhandler/ignoresEmptyDecorator.py
@@ -1,0 +1,2 @@
+@
+def decorated<caret>