+++ testData/typedvalidatormethodhandler/ignoresUnrelatedDecorator.py
@@ -1,0 +1,6 @@
+def unrelated_decorator(function):
+    return function
+
+
+@unrelated_decorator
+def decorated<caret>