+++ testData/typedvalidatormethodhandler/insertsClsForV1Validator_after.py
@@ -1,0 +1,6 @@
+from pydantic import validator
+
+
+class Model:
+    @validator("name")
+    def validate_name(cls, <caret>):