+++ testSrc/com/koxudaxi/pydantic/PydanticCompletionV18Test.kt
@@ -60,8 +60,7 @@ open class PydanticCompletionV18Test : PydanticTestCase(version = "v18") {
                 Pair("ET", "null"))
         )
     }
-    // TODO: Override init keyword argument completion broken in PyCharm 2025.2
-    fun _disabled_testOverrideInitKeywordArgument() {
+    fun testOverrideInitKeywordArgument() {
         doFieldTest(
             listOf()
         )
@@ -113,5 +112,40 @@ open class PydanticCompletionV18Test : PydanticTestCase(version = "v18") {
             .filterNot { excludes.contains(it) }
             .toSet()
         assertTrue(actual.containsAll(setOf("id", "name", "secret_name", "age")))
+    }
+
+    fun testGenericInheritedConcreteField() {
+        doFieldTest(
+            listOf(
+                Pair("a", "int A"),
+                Pair("b", "str B"),
+            )
+        )
+    }
+
+    fun testGenericSubscribedField() {
+        doFieldTest(
+            listOf(
+                Pair("a", "str A"),
+                Pair("b", "int B"),
+            )
+        )
+    }
+
+    fun testGenericBoundField() {
+        doFieldTest(
+            listOf(
+                Pair("a", "int A"),
+                Pair("b", "str B"),
+            )
+        )
+    }
+
+    fun testGenericUnsubscribedField() {
+        doFieldTest(
+            listOf(
+                Pair("a", "AT A"),
+            )
+        )
     }
 }