+++ testSrc/com/koxudaxi/pydantic/PydanticInspectionIncludeExcludeUpdateTest.kt
@@ -1,0 +1,14 @@
+package com.koxudaxi.pydantic
+
+import com.jetbrains.python.inspections.PyInspection
+import kotlin.reflect.KClass
+
+class PydanticInspectionIncludeExcludeUpdateTest : PydanticInspectionBase() {
+
+    @Suppress("UNCHECKED_CAST")
+    override val inspectionClass: KClass<PyInspection> = PydanticInspection::class as KClass<PyInspection>
+
+    fun testIncludeExcludeUpdate() {
+        doTest()
+    }
+}