+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## [Unreleased]
 
+## [0.4.27] - 2026-07-23
+
+- Fix internal API usage [[#1183](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1183)]
+
+## [0.4.26] - 2026-07-20
+
+- Support PyCharm 2026.2 [[#1181](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1181)]
+- Fix crash while typing Python functions [[#1180](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1180)]
+
+## [0.4.25] - 2026-07-01
+
+- Better support for accessing class attributes on table SQLModel models [[#1173](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1173)]
+- Support PyCharm 2026.1.3 [[#1174](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1174)]
+
+## [0.4.24] - 2026-02-26
+
+- Support PyCharm 2026.1 EAP [[#1162](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1162)]
+- feat: Validate model field names in exclude, include and update arguments [[#1147](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1147)]
+- Update Qodana action to v2025.3 and fix linter config [[#1161](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1161)]
+
 ## [0.4.23] - 2025-12-15
 
 - Fix: Enhance SQLModel class completion and implement SDK cache invalidation [[#1137](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/1137)]
@@ -595,7 +615,11 @@
 - Support ellipsis(...) in fields [[#34](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull//34)]
 - Support Schema in fields [[#31](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull//31)]
 
-[Unreleased]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.23...HEAD
+[Unreleased]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.27...HEAD
+[0.4.27]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.26...v0.4.27
+[0.4.26]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.25...v0.4.26
+[0.4.25]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.24...v0.4.25
+[0.4.24]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.23...v0.4.24
 [0.4.23]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.22...v0.4.23
 [0.4.22]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.21...v0.4.22
 [0.4.21]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.20...v0.4.21